backend/tflite_service/requirements.txt
# TFLite service for v2 model. Use Python 3.11 on Windows: py -3.11 -m venv .venv
flask>=2.0.0
Pillow>=9.0.0
numpy>=1.20.0
# v2 model was converted with TF 2.17+ (FULLY_CONNECTED op v12); need TF 2.17+ to load it
tensorflow>=2.17.0
